+++ packages/core/src/vuePlugins/vuex.ts
@@ -10,7 +10,12 @@ import {
 } from '@repositoryname/vuex-generators'
 import noop from '@repositoryname/noop'
 import i18next from 'i18next'
-import { PluginContainer, CoreState, PolarError } from '@polar/lib-custom-types'
+import {
+  CoreState,
+  MoveHandleProperties,
+  PluginContainer,
+  PolarError,
+} from '@polar/lib-custom-types'
 import { Interaction } from 'ol/interaction'
 import { Feature, Map } from 'ol'
 import { CapabilitiesModule } from '../storeModules/capabilities'
@@ -37,6 +42,7 @@ const devMode = import.meta.env.DEV
 
 let map: null | Map = null
 let hovered: null | Feature = null
+let moveHandle: MoveHandleProperties | null = null
 let selected: null | Feature = null
 let components = []
 let interactions: Interaction[] = []
@@ -73,6 +79,7 @@ const getInitialState = (): CoreState => ({
   components: 1,
   center: null,
   hovered: 1,
+  moveHandle: 1,
   selected: 1,
   zoomLevel: 0,
   // TODO: Add default values for epsg, layers, namedProjections, options and remove @ts-ignore for configuration
@@ -106,6 +113,10 @@ const store = new Store({
     map: (state) => {
       noop(state.map)
       return map
+    },
+    moveHandle: (state) => {
+      noop(state.moveHandle)
+      return moveHandle
     },
     hovered: (state) => {
       noop(state.hovered)
@@ -150,6 +161,10 @@ const store = new Store({
     setHovered: (state, payload) => {
       hovered = payload
       state.hovered = state.hovered + 1
+    },
+    setMoveHandle: (state, payload: MoveHandleProperties | null) => {
+      moveHandle = payload
+      state.moveHandle += 1
     },
     setSelected: (state, payload) => {
       selected = payload