+++ packages/plugins/Gfi/src/components/Gfi.vue
@@ -1,49 +1,45 @@
 <template>
-  <div
-    v-if="renderUi"
-    id="polar-plugin-gfi"
-    :class="{
-      'polar-plugin-gfi-move-handle': renderMoveHandle,
-      'polar-plugin-gfi-fullscreen': hasWindowSize,
-    }"
-  >
+  <div v-if="renderUi" id="polar-plugin-gfi">
     <v-card v-if="!windowLayerKeysActive">
       <v-card-text>{{ $t('common:plugins.gfi.noActiveLayer') }}</v-card-text>
     </v-card>
-    <MoveHandle
-      v-else-if="renderMoveHandle"
-      :min-height="0.1"
-      :max-height="maxMobileHeight"
-    >
-      <component :is="contentComponent" v-bind="contentProps"></component>
-    </MoveHandle>
-    <component :is="contentComponent" v-else v-bind="contentProps"></component>
+    <component
+      :is="contentComponent"
+      v-else-if="!renderMoveHandle"
+      v-bind="contentProps"
+    />
   </div>
 </template>
 
 <script lang="ts">
+import compare from 'just-compare'
+import { t } from 'i18next'
 import Vue from 'vue'
+import { mapActions, mapGetters, mapMutations } from 'vuex'
 import { GeoJsonProperties } from 'geojson'
-import { mapGetters } from 'vuex'
-import { MoveHandle } from '@polar/components'
+import {
+  MoveHandleActionButton,
+  MoveHandleProperties,
+} from '@polar/lib-custom-types'
 import Feature from './Feature.vue'
 import List from './List.vue'
 
 export default Vue.extend({
   name: 'GfiPlugin',
-  components: {
-    MoveHandle,
-  },
-  data: () => ({ clientWidth: 0, maxMobileHeight: 1 }),
+  data: () => ({ clientWidth: 0 }),
   computed: {
-    ...mapGetters(['hasSmallWidth', 'hasWindowSize']),
+    ...mapGetters(['moveHandle']),
     ...mapGetters('plugin/gfi', [
+      'actionButton',
       'exportPropertyLayerKeys',
+      'gfiContentComponent',
+      'gfiConfiguration',
+      'renderMoveHandle',
+      'renderType',
+      'showList',
+      'visibleWindowFeatureIndex',
       'windowFeatures',
       'windowLayerKeysActive',
-      'visibleWindowFeatureIndex',
-      'gfiContentComponent',
-      'showList',
     ]),
     contentComponent(): Vue {
       return this.showList ? List : this.gfiContentComponent || Feature
@@ -84,34 +80,79 @@ export default Vue.extend({
     renderUi(): boolean {
       return this.windowFeatures.length > 0 || this.showList
     },
-    renderMoveHandle(): boolean {
-      return this.hasWindowSize && this.hasSmallWidth
-    },
     /** only show switch buttons if multiple property sets are available */
     showSwitchButtons(): boolean {
       return this.windowFeatures.length > 1
     },
   },
+  watch: {
+    actionButton(
+      newButton: MoveHandleActionButton | null,
+      oldButton: MoveHandleActionButton | null
+    ) {
+      if (this.windowFeatures.length && !compare(newButton, oldButton)) {
+        const moveHandleProperties: MoveHandleProperties = {
+          closeLabel: t('plugins.gfi.header.close'),
+          closeFunction: this.closeWindow,
+          component: this.contentComponent,
+          props: this.contentProps,
+          plugin: this.renderType === 'independent' ? 'gfi' : 'iconMenu',
+        }
+        if (newButton !== null) {
+          moveHandleProperties.actionButton = newButton
+        }
+        this.setMoveHandle(moveHandleProperties)
+      }
+    },
+    windowFeatures(features: GeoJsonProperties[]) {
+      if (features.length) {
+        const moveHandleProperties: MoveHandleProperties = {
+          closeLabel: t('plugins.gfi.header.close'),
+          closeFunction: this.closeWindow,
+          component: this.contentComponent,
+          props: this.contentProps,
+          plugin: this.renderType === 'independent' ? 'gfi' : 'iconMenu',
+        }
+        if (this.actionButton !== null) {
+          moveHandleProperties.actionButton = this.actionButton
+        }
+        this.setMoveHandle(moveHandleProperties)
+      } else if (
+        this.moveHandle !== null &&
+        this.moveHandle.component === this.contentComponent
+      ) {
+        this.setMoveHandle(null)
+      }
+    },
+  },
   mounted() {
     window.addEventListener('resize', () => {
       this.updateClientWidth()
-      this.updateMaxMobileHeight()
     })
     this.updateClientWidth()
   },
-  updated() {
-    this.$nextTick(this.updateMaxMobileHeight)
-  },
   methods: {
+    ...mapMutations(['setMoveHandle']),
+    ...mapActions('plugin/gfi', ['close']),
+    closeWindow() {
+      this.close()
+      // The list view is currently only implemented if the gfi is rendered as part of the iconMenu.
+      // TODO: Finding a different solution may be a task to be tackled in the future
+      if (
+        this.gfiConfiguration.featureList &&
+        this.$store.hasModule(['plugin', 'iconMenu']) &&
+        this.$store.getters['plugin/iconMenu/open'] !== null
+      ) {
+        this.$store.dispatch(
+          'plugin/iconMenu/openInMoveHandle',
+          this.$store.getters['plugin/iconMenu/menus'].findIndex(
+            ({ id }) => id === 'gfi'
+          )
+        )
+      }
+    },
     updateClientWidth() {
       this.clientWidth = this.$root.$el.clientWidth
-    },
-    updateMaxMobileHeight() {
-      // Only update if the element is actually rendered
-      if (this.$el.clientHeight) {
-        this.maxMobileHeight =
-          this.$el.clientHeight / this.$root.$el.clientHeight
-      }
     },
   },
 })
@@ -133,23 +174,5 @@ export default Vue.extend({
       overflow-x: clip;
     }
   }
-}
-
-.polar-plugin-gfi-fullscreen {
-  @media only screen and (max-width: 768px) {
-    position: fixed;
-    /* NOTE: If the element is supposed to be lower, a corresponding min-height has to be given to MoveHandle;
-    *  e.g.: top: 95%; -> :min-height="0.5"
-    *  Otherwise it could lead to misleading behaviour. */
-    top: 90%;
-    width: 100%;
-    z-index: 1;
-  }
-}
-
-.polar-plugin-gfi-move-handle {
-  left: 0;
-  right: 0;
-  scrollbar-gutter: unset;
 }
 </style>