+++ packages/plugins/Gfi/src/components/List.vue
@@ -130,9 +130,11 @@ export default Vue.extend({
 </script>
 
 <style lang="scss" scoped>
-#polar-plugin-gfi:not(.polar-plugin-gfi-move-handle) .plugin-gfi-list {
-  /* magic number; leaves minimal space to minimal size AddressSearch in IconMenu mode */
-  max-width: 420px;
+.v-card {
+  @media only screen and (min-width: 769px) {
+    /* magic number; leaves minimal space to minimal size AddressSearch in IconMenu mode */
+    max-width: 420px;
+  }
 }
 
 .v-card__title {